requirements.txt
faiss-cpu>=1.8.0
google-genai>=1.15.0
pypdf>=5.0.0
